use merge-emb-cat for int8 since acc issue is fixed in IPEX (#1477)

+++ models/recommendation/pytorch/torchrec_dlrm/jit_trace_able_utils.py
@@ -26,8 +26,8 @@ def unpack(input :KeyedJaggedTensor) -> dict:
     output = {}
     for k, v in input.to_dict().items():
         output[k] = {}
-        output[k]['values'] = v._values.long()
-        output[k]['offsets'] = v._offsets.long()
+        output[k]['values'] = v._values.int()
+        output[k]['offsets'] = v._offsets.int()
     return output
 
 class SparseArchTraceAbleWrapper(nn.Module):
